PS-6307 Float stock testing

+++ Bundles/Oms/src/Spryker/Zed/Oms/Business/Util/ReservationInterface.php
@@ -21,17 +21,17 @@ interface ReservationInterface
     /**
      * @param string $sku
      * @param \Generated\Shared\Transfer\StoreTransfer $storeTransfer
-     * @param int $reservationQuantity
+     * @param float $reservationQuantity
      *
      * @return void
      */
-    public function saveReservation(string $sku, StoreTransfer $storeTransfer, int $reservationQuantity): void;
+    public function saveReservation(string $sku, StoreTransfer $storeTransfer, float $reservationQuantity): void;
 
     /**
      * @param string $sku
      * @param \Generated\Shared\Transfer\StoreTransfer|null $storeTransfer
      *
-     * @return int
+     * @return float
      */
     public function sumReservedProductQuantitiesForSku($sku, ?StoreTransfer $storeTransfer = null);
